Project restructuration, damage types, abilities WIP

+++ Assets/Code/Entities/Items/BaseItem.cs
@@ -18,7 +18,7 @@ public class BaseItem : ScriptableObject {
 	[Range(1, 99)] public int m_maxStackSize;
 
 	[Tooltip("How much this item sells for")]
-	public float m_sellPrice;
+	public int m_sellPrice;
 
 	[Tooltip("The equipment slots this item fits into, none if not an equippable item")]
 	public List<EquipmentSlot> m_equipmentSlots;
